backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
celery>=5.4.0
redis>=5.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
supabase>=2.4.0
requests>=2.32.0
cryptography>=42.0.0
pypdf>=4.2.0
python-docx>=1.1.0
pytest>=8.2.0
